[elastic] Refactor the deps management.
Move the dependency management from 'protocol/elasticserver.go' to
'lsp/elasticserver.go'. 'protocol/elasticser.go' is more like interface
declaration file, in contrast, 'lsp/elasticserver.go' is more like
definition file. Making the 'manageDeps' as an API of 'ElasticServer'
can get cleaner code.
